pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "hecomes"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "scipy",
    "Pillow",
    "flask",
    "markovify",
]

[project.scripts]
hecomes-image = "hecomes.cli.image:main"
hecomes-video = "hecomes.cli.video:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"hecomes" = ["data/*"]
